Switch to a different fix all provider.
