+++ src/Features/Core/Portable/ConflictMarkerResolution/ConflictMarkerFixAllProvider.cs
@@ -7,21 +7,24 @@
 using System.Collections.Immutable;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.Shared.CodeFixes;
+using Microsoft.CodeAnalysis.CodeStyle;
 
 namespace Microsoft.CodeAnalysis.ConflictMarkerResolution
 {
     internal abstract partial class AbstractResolveConflictMarkerCodeFixProvider
     {
-        private class ConflictMarkerFixAllProvider : AbstractConcurrentFixAllProvider
+        private class ConflictMarkerFixAllProvider : DocumentBasedFixAllProvider
         {
             private readonly AbstractResolveConflictMarkerCodeFixProvider _codeFixProvider;
 
             public ConflictMarkerFixAllProvider(AbstractResolveConflictMarkerCodeFixProvider codeFixProvider)
                 => _codeFixProvider = codeFixProvider;
 
-            protected override Task<Document> FixAllAsync(FixAllContext fixAllContext, Document document, ImmutableArray<Diagnostic> filteredDiagnostics)
-                => _codeFixProvider.FixAllAsync(document, filteredDiagnostics, fixAllContext.CodeActionEquivalenceKey, fixAllContext.CancellationToken);
+            protected override string CodeActionTitle
+                => FeaturesResources.Resolve_conflict_markers;
+
+            protected override Task<SyntaxNode> FixAllInDocumentAsync(FixAllContext fixAllContext, Document document, ImmutableArray<Diagnostic> diagnostics)
+                => _codeFixProvider.FixAllAsync(document, diagnostics, fixAllContext.CodeActionEquivalenceKey, fixAllContext.CancellationToken);
         }
     }
 }
