+++ src/Features/Core/Portable/FeaturesResources.resx
@@ -1756,4 +1756,7 @@ This version used in: {2}</value>
   <data name="Implement_explicitly" xml:space="preserve">
     <value>Implement explicitly</value>
   </data>
+  <data name="Resolve_conflict_markers" xml:space="preserve">
+    <value>Resolve conflict markers</value>
+  </data>
 </root>
