+++ src/Features/Core/Portable/xlf/FeaturesResources.cs.xlf
@@ -537,6 +537,11 @@
         <target state="translated">Místo {0} použijte {1}</target>
         <note />
       </trans-unit>
+      <trans-unit id="Resolve_conflict_markers">
+        <source>Resolve conflict markers</source>
+        <target state="new">Resolve conflict markers</target>
+        <note />
+      </trans-unit>
       <trans-unit id="RudeEdit">
         <source>Rude edit</source>
         <target state="translated">Hrubá úprava</target>
