+++ src/Features/Core/Portable/xlf/FeaturesResources.ja.xlf
@@ -537,6 +537,11 @@
         <target state="translated">'{0}' を '{1}' に置き換える</target>
         <note />
       </trans-unit>
+      <trans-unit id="Resolve_conflict_markers">
+        <source>Resolve conflict markers</source>
+        <target state="new">Resolve conflict markers</target>
+        <note />
+      </trans-unit>
       <trans-unit id="RudeEdit">
         <source>Rude edit</source>
         <target state="translated">Rude 編集</target>
