+++ src/Features/Core/Portable/xlf/FeaturesResources.ko.xlf
@@ -537,6 +537,11 @@
         <target state="translated">'{0}'을(를) '{1}'(으)로 바꾸기</target>
         <note />
       </trans-unit>
+      <trans-unit id="Resolve_conflict_markers">
+        <source>Resolve conflict markers</source>
+        <target state="new">Resolve conflict markers</target>
+        <note />
+      </trans-unit>
       <trans-unit id="RudeEdit">
         <source>Rude edit</source>
         <target state="translated">편집 다시 실행</target>
