+++ src/Features/Core/Portable/xlf/FeaturesResources.zh-Hant.xlf
@@ -537,6 +537,11 @@
         <target state="translated">將 ‘{0}’ 取代為 ‘{1}'</target>
         <note />
       </trans-unit>
+      <trans-unit id="Resolve_conflict_markers">
+        <source>Resolve conflict markers</source>
+        <target state="new">Resolve conflict markers</target>
+        <note />
+      </trans-unit>
       <trans-unit id="RudeEdit">
         <source>Rude edit</source>
         <target state="translated">粗略編輯</target>
